requirement.txt
phidata
groq
python-dotenv
duckduckgo-search
yfinance
openai